MRG: fix more numpy 1.12 indexing errors

More indexing errors found by running the examples.

+++ examples/labs/permutation_test_fakedata.py
@@ -35,11 +35,11 @@ data, vardata, XYZ = make_data()
 P = PT.permutation_test_onesample(data, XYZ)
 
 # clusters definition (height threshold, max diameter)
-c = [(P.random_Tvalues[P.ndraws * (0.95)], None)]
+c = [(P.random_Tvalues[int(P.ndraws * (0.95))], None)]
 
 # regions definition (label vector)
 r = np.ones(data.shape[1], int)
-r[data.shape[1]/2.:] *= 10
+r[int(data.shape[1] / 2.):] *= 10
 voxel_results, cluster_results, region_results = \
                 P.calibrate(nperms=100, clusters=c, regions=[r])
 
@@ -57,7 +57,7 @@ data1, vardata1, data2, vardata2 = (data[:10], vardata[:10], data[10:],
 
 # rfx calibration
 P = PT.permutation_test_twosample(data1, data2, XYZ)
-c = [(P.random_Tvalues[P.ndraws * (0.95)], None)]
+c = [(P.random_Tvalues[int(P.ndraws * (0.95))], None)]
 voxel_results, cluster_results, region_results = P.calibrate(nperms=100,
                                                              clusters=c)
